feat(workspace): Client programs in workspace (#178)
Moves the client programs into the workspace, by using custom build
profiles in the virtual workspace for the client binaries.

This change will allow us to work on the primary client program within
the workspace, without having to have loose crates.

A new crate, `kona-common-proc`, has been added as an extension of
`kona-common` to make clean `main` functions in client programs.

The `#[client_entry(<heap_size>)]` attribute will expand to include a
`_start` symbol if the binary is being compiled to a supported FPVM
target. If not, it will include a `main` symbol, as normal to run on the
native hardware.