fix(node/kurtosis-e2e-test): fixes the versionning for the e2e testing package (#1814)
## Description

This PR fixes the package versions of the go e2e test folder. This fix
ensures that the kurtosis CI job fails if the `kona-node` fails to start
inside kurtosis. This will help preventing breaking the kurtosis
deployment (at least at startup).

Merging this PR means we'll be using the version of the monorepo from my
fork `https://github.com/theochap/optimism`. This will be helpful to add
features to the `devnet-sdk` without waiting for approval inside the
`ethereum-optimism/optimism` repo. Once we stabilize the `devnet-sdk` we
can remove the pointer to the forked version of the node.