feat(node/engine): refactor engine tasks. fix sequencer startup logic (#2388)
## Description

This PR refactors the core logic of the engine task handler. In
particular, it moves duplicate code from the build and the insert unsafe
tasks to the forkchoice task.

Changes:
- Add an associated `Output` type for the `EngineTaskExt` trait which
allows the implementors of this trait to be composed. In particular, the
`ForkchoiceTask` may returns the payload id that was inserted in the EL,
which can be used by the `BuildTask` to consolidate the payload into the
engine state.
- Allows the `ForkchoiceTask` to directly insert newly built payload
attributes into the EL (by adding an `Option<OpAttributeWithParent>`
argument to the `ForkchoiceTask`). This allows to refactors both calls
to the `FCU` inside the `BuildTask` using the `ForkchoiceTask` and
avoids manually handling the forkchoice updates
- Moves the `PayloadStatus` verification logic to the `ForkchoiceTask`
to avoid having to perform the same logic in both the `BuildTask` and
the `InsertUnsafeTask`, which allows to fix the sequencer kickstart
issue described in #2372

Close #2372 

## Potential follow-ups

- Remove the clones inside the engine tasks
- Refactor the `new_payload` logic into an engine task to remove
duplicate code in the `InsertUnsafe` and `Build`Tasks